+++ eventlog-live/CHANGELOG.md
@@ -9,6 +9,7 @@
 - **BREAKING**: Move core machines to their own module.
 - **BREAKING**: Use ANSI colors for logs, if supported.
 - Support reading the eventlog from stdin.
+- Support reading the eventlog from file.
 - **BREAKING**: Rename `EventlogSocket` and related functions to `EventlogSource`.
 
 ### 0.1.0.1